complete qr
@RequiresApi(Build.VERSION_CODES.S) 제거 예정
BottomSheetQrScaffold 컴포넌트로 변경
NullableExtension util 추가
패키지 구조 변경
