+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
         android:theme="@style/Theme.Attendance"
         tools:targetApi="31">
         <activity
-            android:name=".ui.login.LoginProcessActivity"
+            android:name=".feature.login.LoginProcessActivity"
             android:exported="true"
             android:theme="@style/Theme.Attendance"
             android:screenOrientation="portrait"
@@ -35,7 +35,7 @@
         </activity>
 
         <activity
-            android:name=".ui.main.MainActivity"
+            android:name=".feature.main.MainActivity"
             android:screenOrientation="portrait"
             tools:ignore="DiscouragedApi,LockedOrientationActivity">
         </activity>
