+++ app/src/main/java/com/ddd/attendance/feature/qr/screen/QrScreen.kt
@@ -1,0 +1,2 @@
+package com.ddd.attendance.feature.qr.screen
+
