--- app/src/main/java/com/ddd/attendance/ui/component/DDDText.kt
@@ -1,32 +1,0 @@
-package com.ddd.attendance.ui.component
-
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.TextUnit
-
-@Composable
-fun DDDText(
-    text: String,
-    modifier: Modifier = Modifier,
-    textStyle: TextStyle = MaterialTheme.typography.bodyMedium,
-    color: Color = Color.Unspecified,
-    fontSize: TextUnit = TextUnit.Unspecified,
-    fontWeight: FontWeight? = null,
-    textAlign: TextAlign? = null,
-) {
-    Text(
-        text = text,
-        modifier = modifier,
-        style = textStyle,
-        color = color,
-        fontSize = fontSize,
-        fontWeight = fontWeight,
-        textAlign = textAlign
-    )
-}
