--- app/src/main/java/com/ddd/attendance/ui/login/LoginProcessViewModel.kt
@@ -1,28 +1,0 @@
-package com.ddd.attendance.ui.login
-
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class LoginProcessViewModel @Inject constructor() : ViewModel() {
-    fun onClickGoogleLogin() {
-
-    }
-
-    fun onClickSignup() {
-
-    }
-
-    fun onClickNextFromUserName() {
-
-    }
-
-    fun onClickNextFromUserJob() {
-
-    }
-
-    fun onClickNextFromUserAffiliation() {
-
-    }
-}
