--- app/src/main/java/com/ddd/attendance/ui/main/MainActivity.kt
@@ -1,26 +1,0 @@
-package com.ddd.attendance.ui.main
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.ui.graphics.toArgb
-import androidx.core.view.WindowCompat
-import com.ddd.attendance.ui.theme.AttendanceTheme
-import com.ddd.attendance.ui.theme.DDD_BLACK
-import com.ddd.attendance.ui.theme.DDD_WHITE
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        window.statusBarColor = DDD_BLACK.toArgb()
-        WindowCompat.getInsetsController(window, window.decorView).isAppearanceLightStatusBars = false
-        setContent {
-            AttendanceTheme {
-                MainScreen()
-            }
-        }
-    }
-}
