--- app/src/main/java/com/ddd/attendance/ui/main/screen/admin/AdminScreen.kt
@@ -1,24 +1,0 @@
-package com.ddd.attendance.ui.main.screen.admin
-
-import androidx.compose.runtime.Composable
-import androidx.navigation.NavController
-import com.ddd.attendance.ui.main.MainViewModel
-
-@Composable
-fun AdminScreen(
-    navController: NavController,
-    viewModel: MainViewModel
-) {
-    Content(
-        onClickBackButton = {
-            navController.popBackStack()
-        }
-    )
-}
-
-@Composable
-private fun Content(
-    onClickBackButton: () -> Unit,
-) {
-
-}
